refactor(chat): place turn runner in channel layer
